# Faster Gif Decoding