… src/ImageSharp/Formats/Gif/PackedField.cs
… tests/ImageSharp.Tests/TestUtilities/TestImageExtensions.cs