### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/SixLabors/ImageSharp/pulls) open
- [x] I have verified that I am following matches the existing coding patterns and practice as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

### Description
Nothing fancy just a bit of a performance improvement reading and writing gifs. 33% faster decoding but no noticeable change with the encoder. 

We've got 250 GC Generation 2 collects per 1k Operations when encoding that I haven't identified the reason for yet. Possibly caused by multiple short writes to the stream.

**Before**

Method | Runtime |     TestImage |       Mean |      Error |   StdDev | Scaled | ScaledSD |   Gen 0 | Allocated |
--------------------- |-------- |-------------- |-----------:|-----------:|---------:|-------:|---------:|--------:|----------:|
'System.Drawing Gif' |    Core | Gif/rings.gif |   440.1 us |   966.4 us | 54.60 us |   1.00 |     0.00 | 51.2695 |  105.3 KB |
'ImageSharp Gif' |    Core | Gif/rings.gif | 1,231.9 us |   594.7 us | 33.60 us |   2.83 |     0.27 |       - |   2.04 KB |
|         |               |            |            |          |        |          |         |           |
'System.Drawing Gif' |     Clr | Gif/rings.gif |   386.2 us |   198.9 us | 11.24 us |   1.00 |     0.00 | 51.2695 | 105.44 KB |
'ImageSharp Gif' |     Clr | Gif/rings.gif | 1,317.2 us | 1,458.3 us | 82.40 us |   3.41 |     0.19 |       - |   2.09 KB |

**After** 

Method | Runtime |     TestImage |     Mean |    Error |   StdDev | Scaled |   Gen 0 | Allocated |
--------------------- |-------- |-------------- |---------:|---------:|---------:|-------:|--------:|----------:|
'System.Drawing Gif' |    Core | Gif/rings.gif | 391.2 us | 32.14 us | 1.816 us |   1.00 | 51.2695 |  105.3 KB |
'ImageSharp Gif' |    Core | Gif/rings.gif | 909.6 us | 83.44 us | 4.714 us |   2.32 |       - |   2.04 KB |
|         |               |          |          |          |        |         |           |
'System.Drawing Gif' |     Clr | Gif/rings.gif | 392.9 us | 25.92 us | 1.464 us |   1.00 | 51.2695 | 105.44 KB |
'ImageSharp Gif' |     Clr | Gif/rings.gif | 935.4 us | 83.61 us | 4.724 us |   2.38 |  0.9766 |   2.09 KB |



<!-- Thanks for contributing to ImageSharp! -->