@iamcarbon 2018-04-13
@JimBobSquarePants -- I had actually started work to eliminate the EndianBinaryWriter entirely from GifEncoder. It's a heavy abstraction that we don't need here. Can / will take a stab in my next PR after this is merged.